Add in submit test job
